buildroot: update to 2018.05-rc1
qemu: update to v2.12.0

The Buildroot update fixes #28 which broke the build on Ubuntu 18.04,
which had error message:

    conflicting types for 'copy_file_range'

We shouldn't normally update to -rc tags, but it fixes the build and I can't
be bothered to bisect further, when 2018.05 comes we will just update again.

The qemu update is kind of random, and I shouldn't have done it at the same
time, but so be it. I did it because I noticed the build was broken while
trying to fix .gitignores on QEMU.