gem5: build with ccache
CXX='ccache c++' does not work, CXX=/my/path/to/ccache/wrapper/c++
does not work, the only thing that works it to put it in your PATH.

I would rather put a wrapper that forwards to ccache in this repo to
avoid the external host dependency:

    #!/usr/bin/env bash
    ccache c++ "$@"

but then that script has to remove its directory from PATH or else
infinite recursion, and I have no patience to code that.

I considered using the ccache from Buildroot, but it does not setup the
nice /usr/lib PATH, so I gave up on that as well.