gem5: remove build variant from output directory
I noticed that it is not very elegant, since there are many other variants
which are not included, e.g. Linux kernel build and random configs.

Also it makes the path longer for something that is not the most common use case.

Users can use run ID instead when they want separate output directories.